backend/requirements.txt
# Dependências principais
Flask==2.3.3
Flask-Cors
Werkzeug==2.3.7

# Processamento de documentos
python-docx==0.8.11
pdfplumber==0.9.0
reportlab==4.0.4
markdown==3.5.1
beautifulsoup4==4.12.2

# Conversão avançada (opcional)
pypandoc
python-magic-bin

# Utilitários
Pillow==10.0.1
lxml==4.9.3
html5lib==1.1

# Segurança e validação
markupsafe==2.1.3

# Dependências do sistema (instalar separadamente se necessário)
# pandoc - https://pandoc.org/installing.html